remove unused properties
